refactored message.dart
