+++ frontend/lib/screens/chat_screen.dart
@@ -1,10 +1,12 @@
 import 'dart:ui';
-
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:frontend/utils/styles.dart';
 import 'package:frontend/widgets/chat_bottom_sheet.dart';
-import 'package:frontend/widgets/message.dart';
+import 'package:frontend/widgets/proof_request_complete.dart';
+import 'package:frontend/widgets/proof_request_widget.dart';
+import 'package:frontend/widgets/basic_message_widget.dart';
+import 'package:frontend/widgets/proof_request_widget_buyer.dart';
 // import 'package:frontend/widgets/message.dart';
 // import '../models/models.dart';
 import '../services/job_service.dart';
