+++ frontend/lib/widgets/abstract_chat_card.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+abstract class AbstractChatCard extends StatelessWidget {
+  final String sentBy;
+  final DateTime timestamp;
+
+  const AbstractChatCard({
+    super.key,
+    required this.sentBy,
+    required this.timestamp,
+  });
+}
